feat: add ability to create a missing file
The new command will display a quick pick and allow selecting a file to be created. It will create folders in between as well.

While adding the feature to create missing files I've organized a bit.

Also I've updated the switch file quickpick to not display the path of the workspace.